requirements.txt
alpaca-py>=0.30.0
fastapi>=0.100.0
uvicorn>=0.23.0
numpy>=1.24.0
pandas>=2.0.0
optuna>=3.0.0
matplotlib>=3.7.0
